phenomenon-graph/requirements.txt
openai>=1.0.0
pydantic>=2.0.0
click
pyyaml
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
